Fix styles and components
